overlay: Use buffer_get_size API instead of FRAME_SIZE macro
When a GstVideoFrame is created, from a GstBuffer, its size field is based
on the GstVideoInfo extarcted from the negotiated GstCaps. The problem is
that this way the size is based on the generic calculations which don't always
align with the calculations for QTI based memory allocators.
The actual size of QTI based buffers is filled in the GstBuffer so switching
from using the GST_VIDEO_FRAME_SIZE macro to the gst_buffer_get_size API.

Change-Id: I56603f59410542a138818420e81abcebe62b497b