gst-plugin-hexagon-nn: remove duplicated source and headers
Hexagon SKEL library is different for each version of hexagon.
Pre-compiled SKEL library used to be part of this repository and
because of that there was separate sub folders for each version.
But STUB library is the same regardless of the Hexagon version.
Because of that there is no need of separate sub folders for
each version of Hexagon. This Patch remove version sub folders
and move common SKEL source and header files in common folder.

Change-Id: I54e5428f52184a4d4d82da55dd7ce15b99292cba